+++ src/main/kotlin/me/rafaelldi/aspire/services/ResourceListener.kt
@@ -1,0 +1,13 @@
+package me.rafaelldi.aspire.services
+
+import com.intellij.util.messages.Topic
+
+interface ResourceListener {
+    companion object {
+        @Topic.ProjectLevel
+        val TOPIC = Topic.create("Aspire Resource Listener", ResourceListener::class.java)
+    }
+
+    fun resourceCreated(resource: AspireResourceService)
+    fun resourceUpdated(resource: AspireResourceService)
+}