+++ src/main/kotlin/me/rafaelldi/aspire/settings/AspireSettingsState.kt
@@ -4,5 +4,6 @@ import com.intellij.openapi.components.BaseState
 
 class AspireSettingsState : BaseState() {
     var checkForNewVersions by property(true)
+    var connectToDatabase by property(true)
     var collectTelemetry by property(false)
 }